Recompile Webpacker in dev when app/views changes
This fixes the issue where Webpacker would not recompile assets in
development when compiling inline during a Rails request (without the
dev server). By monitoring app/views for changes, we can trigger a
Webpacker compile on the next request and update the purged classes.

This does not affect running the webpack dev server.