🎉 Per Unit Subscriptions (#587)
Adds a `charge_per_unit` flag to Plans. This is used for marking a Plan that charges per unit (like $10/user/mo).

To automatically update the subscription quantity when associated records are created and destroyed, you can add `  updates_subscription_quantity` to your model. This will add after commit callbacks to grab the latest quantity and update the subscription.

This is included by default in the AccountUser model to handle $/user/mo pricing out of the box.

Check out the `UpdatesSubscriptionQuantity` concern for more details. 👍