Use application name for invitation emails
Using the sender's name for the invitation email can cause email clients
to display a spoofing warning.

For example, "Chris Oliver <support@example.com>" could trigger this.
Instead, we'll use the application name and keep the sender's name in
the subject line.

We could also use "Chris Oliver from MyApp" as the email name but I'm
not sure if that would trigger warnings so this will be the safer
option.