Update to action_text-trix gem
This allows Trix to be updated independently from Rails for security updates and includes both the JavaScript and CSS.

Previously, we were pinning Trix with importmaps but this only vendors the JavaScript.

This can be removed with the next Rails release most liekly since ActionText in Rails main depends upon this gem. However, the new editor Jorge is working on may end up as the new default so we'll have to wait until that is released.